Remove VTK-m includes and libs from test compile
All the VTK-m include and libraries should be hidden from code outward
facing of the vtkAcceleratorsVTKm library. Besides, the VTK-m
accelerator tests are not called as a subdirectory from the
vtkAcceleratorsVTKm code, so the CMake variables specifying the include
directories and libraries are not set.

If it is later decided that code using the accelerators library also
needs to directly compile VTK-m code, it should be specified as PUBLIC
includes and libraries to the vtkAcceleratorsVTKm library.