Fix interactionState computation in SeedWidget
When a context menu is displayed over a VTK render window when the mouse was over a seed widget, the VTK window does not get MouseMove events anymore.
If the user clicks outside the context menu to close it, the VTK render window receives the click event.

vtkSeedWidget ignored the click position and assumed that the mouse pointer was still at the same position of the last MouseMove event (and made the context menu appear again, because it sensed it as a click on a seed).

Fixed by forcing update of interactionState using current mouse position in vtkSeedWidget::AddPointAction, the same way as it is already done in vtkSeedWidget::MoveAction.