update shadow maps to use exponential functions
This topic changes the shadow map passes in VTK to use an exponential
function which behaves a bit better and is faster.

also fix a bug where for directional lights the parallel
scale was double what it should be.

Also add in the capability to support subsurface scattering
by compouting a thickness value in the lighting calc.