Add an IgnoreFileTime option to vtkExodusIIReader
When on, this option ignores the time values assigned to each time
step in the file. This can be useful for Exodus files where different
time steps are overloaded to represent different aspects of a data set
rather than the data set at different time values.

A test for this feature is also included.