Add the moment invariants module.
This module provides classes and filters for rotation invariant pattern
detection in 2D and 3D scalar-, vector-, and matrix-valued datasets.
vtkComputeMoments computes the moments. Its output together with the
pattern go into vtkMomentInvariants, which puts out a field containig
the similarity of the pattern to each point.  Detailed information on
the theory can be found in Bujack and Hagen
"Moment Invariants for Multi-Dimensional Data":
http://www.informatik.uni-leipzig.de/~bujack/2017TensorDagstuhl.pdf.
A detailed description of the functionality of the module can be
found here:
http://www.informatik.uni-leipzig.de/~bujack/2017MomentInvariantsModule.pdf