Fix retrieval of increments based on data array
The method GetIncrements(vtkDataArray* scalars, vtkIdType inc[3])
takes a data array that supplies the number of components used in the
increments calculation. However, this method was invoking
ComputeIncrements(vtkIdType inc[3]), ignoring the data array that was
passed in and using the active scalar array in the vtkImageDAta
instead if it was set. This is incorrect, and if an active scalar
array was not set but SetInputArrayToProcess() was, it would lead to
erroneous error messages.