Relax requirements of connectivity filter running in parallel
The parallel connectivity filter would incorrectly return an error
when all data resides on a single process and no ghost points are
available. In this particular case, no ghost points are required
to compute the correct solution, so execution is delegated to
the serial superclass.

In addition, when no ghost points are present in the input, we
just warn that results may not be correct but let the algorithm
proceed.