Fix issue #17211: crash on pick of composite data
vtkCellPicker::IntersectActorWithLine called on composite data
may use incorrect minCellId and minSubId values to determine
picked cell, because vtkCellPicker::IntersectDataSetWithLine
uses th supplied value for parametric distance to determine
if minCellId and minSubId should be updated.

The fix is to pass local variables to
vtkCellPicker::IntersectDataSetWitLine, and to update minCellId
and minSubId if and only if
vtkCellPicker::IntersectDataSetWithLine returns true.