New filter to compute volume and area properties
This filter computers area and volume properties on potentially
multiple objects defined by a polygonal mesh. Objects are identified
by connected traversal; consistent ordering of polygons is detected
and accounted for; and validity of objects (in terms of being closed and
manifold) is detected.

Note: volume calculations are improved by using a nearby point (rather than
the origin) to computed the signed volume.