New method to quickly pick/select points
Analogous to vtkStaticCellLocator, the
vtkStaticPointLocator::IntersectWithLine method implements a fast method
to pick points within a tolerance along a line. The point with a
specified tolerance and closest to the ray start is returned.

Note that finding the points closest to a line means testing some neighboring
bins as well as bins that contain the line.