rework the mtl file handling for obj/mtl files
also add a new exporter designed to work with
obj file although it will work with any scene
in vtk.

The new exporter takes all the actors in a scene
and stores them as a single vtp file. It also takes
all the texturs and combines them into a single texture.
The exporter is designed to take OBJ files that may
consist of mulitple parts each with their own texture
and convert them into a singlke vtp file so that they are
easier to work with in ParaView/etc.