A new test re-creating old Octree case with HyperTreeGrid
This commit does the following:
1. Adds a test to replace the now-retired TestClipHyperOctree.
NB: A better naming is used because the old one was confusing
(this case produces binary trees and quadtrees in addition to
genuine octrees). This also exercises the 1D geometry code
that was not tested before.
2. Includes the corresponding baseline image.
3. Fixed the Orientation variable of the HTG geometry filter,
which was not used correctly in the 1D case, when shifting
the second edge endpoint, with respect to the cursor origin.
4. In 1D also,  the nput data values were not properly passed
to the output edges
5. Made some cosmetic changes to the 3D ternary HTG clipping
test for consistency.