Fixing compiler warnings and an error
The compiler error was that some compilers (Apple Clang 9.0.0.9000039
for a cross-compilation on a Mac for Android for example) don't
recognize strtof() as part of the C++11 standard. The warnings were
efficiency warnings and (in my opinion) an incorrect warning for
"Ineffective call of function 'empty()'. Did you intend to call
'clear()' instead?".

Fixes #17226.