first working version of file format for HyperTreeGrid
What I've done so far in the implementation is write a set of .vhg files
from the original structure created by the adaptor and then reading that
back in with my reader and writing a second set of  .vhg files from that
structure and making sure they match. Also I took those two structures
and did a convert to unstructured grid as it appears in the Filters code
and produced two sets of .vtu files which also match.

I treat the descriptor as a vtkCharArray for now but it will be an easy
change to vtkBitArray.  I did not bother with the MaterialMask or
PureMaterialMask at all figuring it would be better to wait for the bit
array implementation.  So all I put in the XML is the series of
attributes describing size of the grid and coordinates (with place
holders for origin and scale once we get the UniformHyperTreeGrid) and
then the MaterialMaskIndex giving the trees on a particular processor
and the matching descriptor to build those trees.