Improving the vtkGradientFilter error message for a missing array.
If the vtkGradientFilter is used over a composite dataset with
the selected array not being available on all non-empty composite
dataset blocks the vtkGradientFilter will return an error for
those blocks. The error message when that happens is now improved
to let users know that the error may be the result of the missing
arrays.

Addresses paraview/paraview#17551.