Added threaded random sequence generation
Added a new class vtkRandomPool which creates a sequence of
random numbers in an array. The processing for large numbers
is threaded. The pool can be used in multithreaded applications
where a random sequence can be precalculated, and the then
deterministically accessed to provide reproducible results. Pools
can be created using difference subclasses of vtkRandomSequence.
By default, vtkMersenneTwister is used.