remove legacy method signature for multitexture
Marking one signature of a method as legacy when it
is overridden is a mess as it results in either a
partial override or a deprecated override method both
of which result in a warning so just removing that sig.
The method has been a nop for a few releases.