add ability to share data between render windows
This topic adds a capability to share some data between
render windows. Basically it is implemented shared context
data in OpenGL such as with wglShareLists or the similar
functionality on glX or Cocoa. Right now only the VBO
cache makes use of the shared space but in the future shader
programs, textures, etc could also be shared.