Use a tolerance for detecting zero width lengths.
A customer reported a bug in the cell locator crashing from not
being able to allocate sufficient memory. The dimensions of the dataset
were (2e-14, 1000, 500), and because of the 2e-14 length
vtkBoundingBox::ComputeDivisions was creating roughly 11M bins to
hold 60 cells :)

Fixed by considering any lengths less than 0.1% of the average dataset
width to be effectively zero.