Fix picking through a disabled widget with picking manager on
With picking manager on, it was impossible to pick a widget if a disabled
widget was in front of it.
To fix this, we have:
  - In the old style widgets (derived only from vtkInteractorObserver), we
went through all the implementation and added the unregistration/registration
of the pickers in the SetEnabled() method.
 - In new style widgets (derived from vtkAbstractWidget and
vtkWidgetRepresentation) the method RegisterPickers and UnRegisterPickers
have been moved into the representation public API to be able to be called
by the widget's SetEnabled() method. This allows the widget to register and
unregisters the pickers as necessary when enabled and disabled.

In both cases we also implemented the SetPickingManaged method as previously
the PickingManaged property was never used anywhere in the code base.
Similarly, the PickersModified() method was removed as its purpose seemed
redundant to SetPickingManaged().

The test TestPickingManagerSeedWidget2 was added to demonstrate/test the
picking behind a disabled widget.

For more background information, see https://issues.slicer.org/view.php?id=3808

Co-Authored-by: Ken Martin <ken.martin@kitware.com>
Co-Authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>
Thanks: Steve Pieper <pieper@bwh.harvard.edu>