Fixing ghost cells generator filter bug and efficiency improvements
When some processes had ghost cells and cells and other processes
had no cells (and thus no ghost cells were needed or kept) the
processes with ghost cells would return early from the filter while
the other processes did not. This resulted in a hang. Also improved
the logic on when to compute ghost cells and when to use existing
ghost cells.