remove the oculus module as it is not maintained
Due to changes in the oculus SDK build/link process
this module has not compiled for a long time.

It was modified months ago to be an error should you
even try to turn this module on.

The code has been moved to a public repo should someone
want to continue work on it. It can be found here

https://gitlab.kitware.com/ken-martin/vtk-oculus