Improve vtkPConnectivityFilter performance
Refactor the class to drastically reduce the data exchanged in
vtkPConnectivityFilter. Now, only points that may be potentially
needed by neighboring ranks are exchanged. A rank's neighbors are
determined as ranks whose data axis-aligned bounding box overlaps the
current ranks bounding box. Non-blocking point-to-point communication
routines are used to exchange points and RegionID arrays. This filter
uses points received from its neighbors to correlate RegionIDs on
different ranks.

In addition, this class no longer requires ghost points or cells. They
were not really needed in the first place, so the warning about them
being needed has been removed. The filter still works if ghost points
are present, however, so they need not be removed. To make this work,
we send not just ghost points to neighbors, but non-ghost points as
well. Non-ghost points work just as well in the RegionID correlation,
hence the removal of the requirement for ghost points in the data set.

Also, introduced the use of PIMPL idiom to help split the algorithm
into smaller pieces.