Handle mismatch is chosen and available arrays.
vtkXMLDataReader had asserts that checked for extact match between
arrays enabled using the array selection API and then the arrays
actually read (present) in the file. This caused issues when reading
composite datasets with partial arrays following the change in
56f1f110 (see paraview/paraview#18189).

Updating the code in vtkXMLDataReader so that it can handle a mismatch
in fields. The limitation was unreasonable anyways.