Fix For Wrong Field Name Passed To PointAverage Filter
A bug causes the vtkmGradientFilter class to fall back to vtkGradientFilter
due to an exception thrown by PointAverage filter.
This is fixed by passing the correct field name to
PointAverage filter.
The issue was reported as a segmentation fault (issue #213).
However, the segfault was not reproducible since the
code has changed since then.