Fix mis-coloring when mixing lines and polys
Previously the code assumed that the full colors array would match
up with either the lines or the polys, but this is not the case when
the polydata has both lines *and* polys.  We have to iterate over
all the cells twice when rendering in order to make sure the color
array is indexed properly for each cell.