Fix vtkLocationSelector
Previously, vtkLocatorSelector attempted to find points by iterating
over each point in the data set and then checking whether one of the
locations of interest is within epsilon of the point. If this was the
case, the point would be marked as closest to a location of interest.
This criteria is not correct for finding the closest point and could
lead to all points spuriously being selected by, for instance, setting
epison large enough to contain the entire data set.

This patch fixes the algorithm error by setting up a point locator for
the data set and using it to determine the point closest to each
location. This matches the previous behavior in
vtkExtractSelectedLocations.