Reduce amount of point data exchanged in vtkPConnectivityFilter
Points can be represented by not just doubles but floats or any
type. Rather than cast points to doubles and exchange double arrays
among ranks, change the implementation to handle the different point
type arrays using the vtkDataArrayDispatcher mechanism. This required
a significant restructuring of the code, but left the algorithm
basically the same.