Ensure that the glyphNormalMatrix is used in shader
Before trying to add the glyphNormalMatrix as an attribute for the
vertex shader in vtkOpenGLGlyphHelper, make sure it is used in the
program. This prevents error messages about the attribute not being
found in the program.

Modified existing test that removes all lights from the renderer to
exercise the former error condition and verify that it is fixed.