Clip Polyhedra Crashes for N-faced Elements
When clipping a dataset that contains n-faced elements (polyhedra),
the program crashes. This is due to use of point locator that merges
similar points and causes mismatch in number of points and point-Ids
stored in the clipped cells.
The issue was fixed by updating the cells with new ids where the
point locator is used in vtkPolyhedron::Clip function.
Also, added a test case for clipping a polyhedra dataset.