vtkTessellatorFilter: Warn and skip data arrays that exceed size limit
vtkTessellatorFilter uses vtkStreamingTessellator to tessellate datasets.
In vtkStreamingTessellator, a compile-time limit
"vtkStreamingTessellator::MaxFieldSize" restricts the number of field values
that can be passed through the filter. When the aggregate number of field
data components exceeded this limit, vtkEdgeSubdivisionCriterion would issue
an error message, but the algorithm would proceed and then crash. This update
modifies vtkEdgeSubdivisionCriterion::PassField to return a failure flag when
the size limit has been reached (as the documentation for the method states it
should). It also modifies vtkTessellatorFilter to skip attempting to copy data
arrays that exceed this limit, instead warning that the arrays will be omitted
from the result.