rework offscreen rendering to be centralized
Move a lot of code into OpenGL and simplify. Introduce
ShowWindow and update UseOffScreenBuffers. These two booleans can
work together to handle all four cases. You can...

ShowWindow = 0 UseOffScreenBuffers = 0
You are using a window to render but the window
is not current shown on the display

ShowWindow = 1 UseOffScreenBuffers = 0
You are rendering into a window and the window
is shown on the screen. The default.

ShowWindow = 0 UseOffScreenBuffers = 1
You are rendering into a framebuffer offscreen
and the window used for the opengl context is not
shown on the display.

ShowWindow = 1 UseOffScreenBuffers = 1
You are rendering into a framebuffer offscreen
and there is a window that is also visible with it's
own content.

For backwards compatability OffScreenRendering now
just sets ShowWindow(false) and UseOffScreenBuffers(true)

Other more specific approaches, methods have been removed.

Mapped has been fixed a bit in some backends to properly
indicate if the window is mapped (as opposed to created)