fix issues with order independent translucency_fixes
Two issues, the depth settings were left in an odd state and
when the default framebuffer was a msaa backed depth buffer
using fixed rep it failed. Added logic to guess float versus
fixed for the depth buffer.

Also cleaned up and simplified a bit as we were using a
texture that we didn't need to use.