SetCellType stores and reuses instantiated cells
SetCellType is typically called on each cell of a vtkDataSet.
When the cell type varies often as may be the case with a
vtkUnstructuredGrid, a new cell type is currently instantiated
each time. This change stores the dynamically allocated cell in
a fixed-length array of cell pointers and reuses it when its type
is requested again.