Add RegionId assignment mode
By default, the mode is UNSPECIFIED, which matches the previous
behavior of how RegionIds are assigned to different connected
components.

Another option, CELL_COUNT_DESCENDING, has also been implemented. In
this mode, RegionIds are assigned to connected components such that
the lowest RegionId is assigned to the component with the largest
number of cells and other RegionIds are assigned in descending order
of cell count.