BUG: vtkOBJImporterr does not handle multiple groups using same material
The original code assumed that each material could only be used by one
actor (group). There is no mention of this restriction in the
Wavefront obj specification. The solution is to use a map of vectors
to map between material and actors.

BUG: vtkOBJExporter fails if vt are present but texture indices not present
Another bug was discovered while addressing the original issue. If
texture coordinates are present via "vt" command but texture indices
are not defined for the faces, the actor was skipped. The solution is
to add a boolean variable that indicates whether texture indices are
defined. If there are no texture indices, do not add TCoords to the
PointData.