Introduce IdType attribute tag in XML format
This adds a IDType attribute to the XML format
in order to identify VTK_ID_TYPE array.

If the type of the dataarray is smaller
than the size of IdType on the vtk build,
it is not considered an error but the resulting
array will not be an vtkIdTypeArray

This adds a test for it.