Added high-performance contouring algorithm
This filter is a specialized, threaded, high-performance
implementation of an isocontouring algorithm for unstructured
grids. In particular, it generates output triangles from 3D linear
cells: vtkTetrahedron, vtkHexahedron, vtkVoxel, vtkWedge, and
vtkPyramid.

The commit includes supporting classes for edge-based point merging
of coincident points, and modification to some VTK cells in order
to extract edge and case tables in support of isocontouring.