Fix locale issue with Legacy VTK readers & writers
Previous implementation is sensible to the locale settings.
This can lead to the creation of VTK files with decimal separators
that, for instance, are not compatible with applications like ParaView
that sets the "C" locale to avoid such conflicts.
This fix set the "C" locale before reading/writing files and restore
the previous value when done.