add logging infrastructure
Introducing a logging framework in VTK via the vtkLogger. It provides
ability for applications based on VTK to generate logs.

vtkDemandDrivenPipeline, vtkStreamingDemandDrivenPipeline now log
various pipeline passes on the algorithm when maximum verbosity log is
being genereted. vtkResourceFileLocator and vtkPythonInterpreter also
use vtkLogger for logging debugging information.

Added a test, TestLogger, to test (and demo) the capabiltiies.

Standard VTK error/warning macros also forward the messages to the
logging framework. To avoid duplicate messages from being posted to the
stdout/stderr from vtkOutputWindow and from vtkLogger, added a new
`display-mode` concept to vtkOutputWindow which allows for a mode to not
post to terminal for error/warning macros when they are being logged.