Improve Context2D labeled isolines
* Fix label orientations under conditions when the viewport is not
square (i.e. when it has aspect ratio far from 1.0).

* Also remove the rendering of the background quad, as that is
already done for us by the DrawString api method, so we were getting
two quads drawn behind each label, and the one we drew manually suffered
from stretching and skewing under the viewport conditions described
above.

* Additionally fix the computation in display coords of the corners
of each label bbox, which was resulting in label overlap testing
not catching too many labels in the same neighborhood.

* Also rename those global structs in the implementation file, as
some compilers (MSVC in some cases) have problems with the fact
that the vtkLabeledContourMapper has global structs with the
same names.

* Finally, remove some code that is not being used.  This code was
initially copied from the 3D labeled contour mapper, and some parts
were left in place at the time even though they are not needed in
a 2D coordinate system.  There are still cleanups that remain to be
done along these lines, but this is a step in the right direction.