OpenGL: add option to say libGL has GLX
Some older implementations of OpenGL have glX symbols included in
libGL. Add an option VTK_OPENGL_LIBGL_HAS_GLX that should be set to ON
if this is the case. Note: newer versions of OpenGL tend to have libGL
and libglx split into separate libraries, in which case this new
option should be OFF. A hint for the appropriate setting of the option
is provided.