Make denoising controls accessible on the renderer
ivars on the viewnodes are generally inaccesible to user code.
Continue to use the information key approach to extend VTK's
primary rendering class API with controls that are specific to
ospray.

Denoiser is hard to test because vtk's fuzzy image comparisons
make them blind to the clarity effect. So, the "N" keypress
in the interactive tests (-I) lets you manually toggle it to
verify it works. Decrease number of samples "p" to make it more
obvious.