Set absolute tolerance instead of tolerance
vtkAppendDataSets was forwarding its tolerance to the Tolerance
property of the vtkCleanPolyData filter when merging was on and the
output is polydata. However, the vtkCleanPolyData filter was set to
use AbsoluteTolerance, which is a different property, and it was using
the default AbsoluteTolerance, leading to severe point merging of far
away points. Fix by setting the AbsoluteTolerance in the clean filter
instead.