Add a generic adios2 reader to read bp format data as images into vtk
This commit adds a reader that can read the arrays in the bp file as
point or cell arrays on a 2d or 3d vtk image. It does not need reader to provide
and schema infomation. It also supports time step and
reading in parallel mode. The minimum version of ADIOS2 requirement is 2.4.0.

The features of this reader is listed as below:

* The output of the reader is a vtkMultiBlockDataset. It contains a
vtkMultiPieceDataSet filled with vtkImageData blocks. The number of blocks
equals to the number of blocks in the adios array.

* It supports arrays with time steps.

* It supports 2D and 3D images.

* The reader can either work in serial or in parllel.

* Based on the number of processes, it will automatically distribute the
work evenly among processes.

* It support row major and column major conversion.