Add ToleranceIsAbsolute option for point merging
Added new option in vtkAppendFilter and vtkAppendDataSets to indicate
that the tolerance is absolute. Previously, the tolerance was always
absolute, hence the default value for this option is 'true'. When it
is false, the Tolerance member variable is treated as a relative
tolerance, meaning it is multiplied by the length of the diagonal of
the dataset's bounding box. This matches the definition of absolute
and relative tolerance in vtkCleanPolyData.