change how edges are handled with coincident settings
The old code for surface with edges used a fixed heuristic
to set the offset params for edges on polygons and strips.
As such it could not be changed without modifying the code.
This patch makes edges use the same settings as lines which
allows apps and users to control the settings.