vtkAlgorithm: progress shift/scale
vtkAlgorithm now supports ability to specify values for shifting and
scaling progress. This enables internal vtkAlgorithm instances to report
progress for part of the work without too much work. Also allows
vtkCompositeDataPipeline to report progress was each iteration over the
composite dataset as a part of the whole rather than having the progress
be from [0, 1.0] for each block in the composite dataset.