vtkModule: support "forward linking"
This feature allows for private dependencies to expose usage
requirements to transitive consumers. The primary use case for this is
for optional linking of libpython where the consuming executable must
link to libpython even though it links to something which private links
to Python. CMake's usage requirements do not support this use case.