VTK now provide a global control to enable CUDA
A limitation of CMake is that enable_language calls most occur
in the top-most file of any target that uses or links to a target
of said language.

This means that for VTK we need to enable CUDA at the top level as
when building statically with python enabled the consumers of
a target that use CUDA is at the top-most level.