vtkModule: remove obsolete comments
When exporting, module names are targets, so the
`_vtk_module_export_properties` comments can be removed. In
`_vtk_module_build`, the check for a module-as-imported-target is fine
because if it is a target, then it could be a real problem. The optional
dependency check is OK because the module will be available by its own
name as a target if it exists which is what we're testing anyways.