Add option to ignore missing arrays
Add an IgnoreMissingArrays flag. When on, if a variable is defined in
the calculator via vtkArrayCalculator::Add*ArrayName() and
vtkArrayCalculator::Add*Variable(), but the field data does not have
an array with the associated name, this filter will not issue an error
and will just silently not produce a result array. When off, the
filter will produce an error (and also not produce a result array).

This option is useful when dealing with partial arrays in
vtkCompositeDataSets. Instead of issuing an error, vtkArrayCalculator
will simply not produce a result if an array for a variable in the
calculator expression does not exist in a block.