again try to fix this test
This is a mess because it seems like the pipeline created
has no indication of what nodes are what level in the
tree. So we end up having to iterate over the last
filters input to count non lowest level nodes and
then use that count when iterating over the mappers
input.