Adding an SoA array class that has a scale factor
For zero-copying arrays for in situ work, this allows the situation
where we want to scale the values that are stored in the memory
by a constant value. This is useful for a simulation code that
has multiple physics/simulation codes working together with
different scalings in each physics/simulation so that when we
get to doing zero-copy of their memory we can scale their field
data into a common scale/unit.