some opengl related cleanups and features
Add ability to import a txture by handle to a texture object

Fix capitalization of ResetGl to be ResetGL

Add SetShowControllerLabels for OpenVR

Fix unused parameter warning in OpenGLRenderWindow

Change state in debug builds to fix a mismatch in state
once detected. This should be safer and should reduce the
number of warnings received when a mismatch occurs.