Fix streaming in vtkImageMarchingCubes
This filter wasn't streaming because the input's whole extent was
updated before vtkImageMarchingCubes::RequestData() was called.
In order for streaming to work properly, the input data must not
be updated before RequestData() updates it.  To state this in
another way, setting a memory limit for the input has no effect
if the whole input is filled before vtkImageMarchingCubes even
executes.

To properly test this filter, the test was modified to use a
streaming reader (vtkImageReader2 instead of vtkVolume16Reader),
and the InputMemoryLimit was reduced.