Fix for wrapping multiple classes per header
The MergeSuperClasses() function for the Python wrappers traces
though the parents of each wrapped class to search for method
overrides and to handle using declarations.  This function had a
bug that caused it to erase the methods from the superclass'
ClassInfo data struct.  As a result, if a class and its superclass
were in the same header and therefore wrapped in the same call to
the vtkWrapPython executable, then superclass methods would not
be wrapped.