fix two issues with probing/finding cells
First issue is that a tolerance squared was being passed
into a function needing distance. Needed a sqrt.
The second issue is that the default computed tolerance for
the proble filter was absurd and wasn't computed at all
just hardcoded to max double which causes numerous overflows
and incorrect behavior. Changed the logic to look at some
actual cells and compute a tolerance from them.