Allow for static builds of vtk to usable by python
Previously the python bindings only worked when VTK was built
dynamically. Do this we need to teach the python module load
code that vtk python modules are all built-in.

The trick is that it is unclear in the Python-C API how to construct
`namespace.module` import that doesn't trigger a search for modules
based on the `.`. So instead at the C++ level we import
"namespace_module" during startup and than the python modules moving
those imports into the correct python module.